Fixes
- Split contract: split amount calculation method
- Removing redundant variables
- Using asynchronous method to get balance
